+++ bootstrap/providers.php
@@ -30,6 +30,7 @@ return [
     Webkul\MagicAI\Providers\MagicAIServiceProvider::class,
     Webkul\Marketing\Providers\MarketingServiceProvider::class,
     Webkul\Notification\Providers\NotificationServiceProvider::class,
+    Webkul\PayU\Providers\PayUServiceProvider::class,
     Webkul\Payment\Providers\PaymentServiceProvider::class,
     Webkul\Paypal\Providers\PaypalServiceProvider::class,
     Webkul\Product\Providers\ProductServiceProvider::class,