+++ packages/Webkul/Admin/src/Config/system.php
@@ -1878,6 +1878,7 @@ return [
                 'name'    => 'sort',
                 'title'   => 'admin::app.configuration.index.sales.payment-methods.sort-order',
                 'type'    => 'select',
+                'default' => 1,
                 'options' => [
                     [
                         'title' => '1',
@@ -1891,6 +1892,12 @@ return [
                     ], [
                         'title' => '4',
                         'value' => 4,
+                    ], [
+                        'title' => '5',
+                        'value' => 5,
+                    ], [
+                        'title' => '6',
+                        'value' => 6,
                     ],
                 ],
             ],
@@ -1983,6 +1990,7 @@ return [
                 'name'    => 'sort',
                 'title'   => 'admin::app.configuration.index.sales.payment-methods.sort-order',
                 'type'    => 'select',
+                'default' => 2,
                 'options' => [
                     [
                         'title' => '1',
@@ -1996,6 +2004,12 @@ return [
                     ], [
                         'title' => '4',
                         'value' => 4,
+                    ], [
+                        'title' => '5',
+                        'value' => 5,
+                    ], [
+                        'title' => '6',
+                        'value' => 6,
                     ],
                 ],
             ],
@@ -2073,6 +2087,7 @@ return [
                 'name'    => 'sort',
                 'title'   => 'admin::app.configuration.index.sales.payment-methods.sort-order',
                 'type'    => 'select',
+                'default' => 3,
                 'options' => [
                     [
                         'title' => '1',
@@ -2086,6 +2101,12 @@ return [
                     ], [
                         'title' => '4',
                         'value' => 4,
+                    ], [
+                        'title' => '5',
+                        'value' => 5,
+                    ], [
+                        'title' => '6',
+                        'value' => 6,
                     ],
                 ],
             ],
@@ -2142,6 +2163,7 @@ return [
                 'name'    => 'sort',
                 'title'   => 'admin::app.configuration.index.sales.payment-methods.sort-order',
                 'type'    => 'select',
+                'default' => 4,
                 'options' => [
                     [
                         'title' => '1',
@@ -2155,6 +2177,12 @@ return [
                     ], [
                         'title' => '4',
                         'value' => 4,
+                    ], [
+                        'title' => '5',
+                        'value' => 5,
+                    ], [
+                        'title' => '6',
+                        'value' => 6,
                     ],
                 ],
             ],
@@ -2241,6 +2269,115 @@ return [
                 'validation'    => 'max:200',
                 'channel_based' => false,
                 'locale_based'  => true,
+            ], [
+                'name'    => 'sort',
+                'title'   => 'admin::app.configuration.index.sales.payment-methods.sort-order',
+                'type'    => 'select',
+                'default' => 5,
+                'options' => [
+                    [
+                        'title' => '1',
+                        'value' => 1,
+                    ], [
+                        'title' => '2',
+                        'value' => 2,
+                    ], [
+                        'title' => '3',
+                        'value' => 3,
+                    ], [
+                        'title' => '4',
+                        'value' => 4,
+                    ], [
+                        'title' => '5',
+                        'value' => 5,
+                    ], [
+                        'title' => '6',
+                        'value' => 6,
+                    ],
+                ],
+            ],
+        ],
+    ], [
+        'key'    => 'sales.payment_methods.payu',
+        'name'   => 'admin::app.configuration.index.sales.payment-methods.payu',
+        'info'   => 'admin::app.configuration.index.sales.payment-methods.payment-method-info',
+        'sort'   => 6,
+        'fields' => [
+            [
+                'name'          => 'active',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.status',
+                'type'          => 'boolean',
+                'channel_based' => true,
+                'locale_based'  => false,
+            ], [
+                'name'          => 'title',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.title',
+                'type'          => 'text',
+                'validation'    => 'required_if:active,1',
+                'channel_based' => true,
+                'locale_based'  => true,
+            ], [
+                'name'          => 'description',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.description',
+                'type'          => 'textarea',
+                'channel_based' => true,
+                'locale_based'  => true,
+            ], [
+                'name'          => 'image',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.logo',
+                'type'          => 'image',
+                'info'          => 'admin::app.configuration.index.sales.payment-methods.logo-information',
+                'channel_based' => false,
+                'locale_based'  => false,
+                'validation'    => 'mimes:bmp,jpeg,jpg,png,webp',
+            ], [
+                'name'          => 'sandbox',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.sandbox',
+                'type'          => 'boolean',
+                'channel_based' => true,
+                'locale_based'  => false,
+            ], [
+                'name'          => 'merchant_key',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.merchant-key',
+                'info'          => 'admin::app.configuration.index.sales.payment-methods.payu-merchant-key-info',
+                'type'          => 'password',
+                'validation'    => 'required_if:active,1',
+                'channel_based' => false,
+                'locale_based'  => false,
+            ], [
+                'name'          => 'merchant_salt',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.merchant-salt',
+                'info'          => 'admin::app.configuration.index.sales.payment-methods.payu-merchant-salt-info',
+                'type'          => 'password',
+                'validation'    => 'required_if:active,1',
+                'channel_based' => false,
+                'locale_based'  => false,
+            ], [
+                'name'    => 'sort',
+                'title'   => 'admin::app.configuration.index.sales.payment-methods.sort-order',
+                'type'    => 'select',
+                'default' => 6,
+                'options' => [
+                    [
+                        'title' => '1',
+                        'value' => 1,
+                    ], [
+                        'title' => '2',
+                        'value' => 2,
+                    ], [
+                        'title' => '3',
+                        'value' => 3,
+                    ], [
+                        'title' => '4',
+                        'value' => 4,
+                    ], [
+                        'title' => '5',
+                        'value' => 5,
+                    ], [
+                        'title' => '6',
+                        'value' => 6,
+                    ],
+                ],
             ],
         ],
     ], [