+++ packages/Webkul/PayU/src/Models/PayUTransactionProxy.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Webkul\PayU\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class PayUTransactionProxy extends ModelProxy {}